TestHideHeavenStem
1. TestHideHeavenStem 文件名更改
2. 方法名修复
